refactor: перенес страницу логина в папку auth/pages
